give in to jukoz' weird view on tidiness
